feat(map): ask where a place is and what it is on the one screen
Adding a place was two screens. The first handed the map over to a pin and
asked only to have it aimed; confirming it put the map away and opened a
full-screen form for the name and the category, with a "Change" link to go
back and move the pin again. The two questions are one question — a shop is a
name at a spot — and splitting them meant the answer to each was invisible
while the other was being given. The map now keeps the top half of the screen
with the pin live on it, and the form takes the bottom half with its submit
button at the foot.

The form is the map's sibling, not something over it

This is the decision the rest of the change falls out of. The map view is no
longer the whole screen: it and everything drawn on it live in a `mapArea`
that takes `flex: 1`, and the form takes the other `flex: 1` beside it.

The alternative was to leave the map at full height under a floating sheet and
push its centre up with `mapPadding`. That was rejected on a platform
difference rather than on taste. On Android `mapPadding` reaches
GoogleMap.setPadding, which does move the camera target into the padded area,
but react-native-maps maps it to `layoutMargins` on iOS, and MKMapView's
`centerCoordinate` stays the geometric centre of its bounds regardless. The
pin would have been drawn where the map looked centred and submitted where it
actually was — a silent, per-platform coordinate bug on the one screen whose
whole job is a coordinate.

Halving the map view instead keeps the existing invariant untouched: the pin
is drawn at 50% of the map view, so where it points is the region's centre,
and there is still nothing to measure or convert. Both platforms hold the
camera target across a resize, so opening the form leaves the map on the spot
it was already showing.

It also had to stop being a native modal. A modal window takes every touch on
the screen, and the map above has to stay pannable for the whole time the form
is open — which is the point of the change.

What that made unnecessary

`locatorBarTop` is gone, and so is the `bottom` prop on the locate-me button
that existed to consume it. Both were there to lift that button clear of the
locator's bar; the button is positioned inside the map's half now, so "the
bottom of the map" already means the top of the form. `PlaceLocator` is the
pin and its hint and nothing else — its Cancel and Next were the two-step
flow's controls, and the form's close and submit are the one-step flow's.

`addStep` collapses from "locating" | "describing" | null to a boolean, and
`pinnedLocation` disappears: there is no moment of freezing a location any
more, only the map's current centre. The attempt counter drops from state to a
plain ref, since the only thing that ever rendered from it was the form's
`key` and the form is now unmounted outright when an attempt ends — which is
what throws away what was typed.

Two behaviours change, and both are forced

The coordinates are on screen beside the map now, and that inverts two rules
that were right when they were not.

It submits the camera the map has settled on rather than the one still in
flight. The old code deliberately read the moving camera: the form was opened
by a tap on Next, and reading the settled region would have placed the pin one
gesture behind wherever the map was heading. Nothing is opened by a tap any
more, and the coordinates are readable on the form throughout — so what is
sent has to be what that row is showing, and a row can only show a camera that
has come to rest.

A failed send no longer clears when the pin moves. That rule was written when
moving the pin meant deliberately leaving the form for the map, which said
plainly that the place was being changed. The map is under the form at all
times now, so the same rule would let an idle nudge wipe a message before it
has been read. It clears on a typed edit, or on the next send.

While a send is out the location row holds the place that was sent rather than
following the map. The request carries the pin as it stood when submit was
tapped and the map is still pannable behind it, so a row that kept following
would be naming a place the request is not going to. This is the same concern
the old form answered by disabling its "Change" link, which has no equivalent
now that there is no link.

Keyboards

Android resizes its own window; iOS does not, so the container is a
KeyboardAvoidingView while a form is open — and only while one is, or the
search modal's own keyboard would resize the map behind it. The two halves go
on splitting whatever height is left evenly, which is deliberate: typing takes
room from the map as well as from the form, so the pin stays in view while the
name of what it is pointing at is being typed. Nothing is focused on open, for
the same reason — a keyboard on arrival would cover the map the form exists to
be filled in alongside.

Tests

add-place-modal.spec.tsx moves with its subject. The tests that drove the
second step through the first are rewritten to drive one screen; the form's
being on screen is now its being mounted at all, since that is how the map
opens and closes it. Two go: "goes back to the map when the pin needs moving"
describes a trip that no longer exists, and "closes the attempt when the answer
lands after the pin went back on the move" lost the state it distinguished —
what it guarded is covered by the success path closing the attempt and by the
abandoned-attempt tests either side of it. New ones cover the pin and the form
appearing together, the form tracking the map as it is panned, the settled
camera being the one submitted, the row holding still during a send, and the
failure surviving a pan.

Every one of those five was mutation-tested rather than trusted: handing the
form `regionRef.current` instead of the settled centre, dropping the held
location, restoring the clear-on-pan effect, re-keying the form on the region,
and mounting the form without the pin each fail the test that claims to
prevent them.

The map component's 186 tests pass, with tsc, eslint and prettier clean. The
full suite is 9224 passing and one failing — a Naira formatting case in
send-confirmation.spec.tsx, which fails identically on a clean checkout of the
parent commit.

Left alone on purpose

The design this was taken from also renames things and reshapes the fields:
"Suggest business" for the title, "Submit request" on the button, a dropdown
where the category chips are, pencils beside the rows. The copy is untouched
because a new key means all 28 locale files, and filling 27 of them with
English is not a decision to make in passing. The fields are untouched because
the ask was the flow. The address row does lose its edit affordance, which is
not a shortfall: the map above it is that control, and it is on screen.

The search bar and the add-place pill still come off the map while a place is
being added, per the decision already recorded there. The reference screenshot
shows the search bar present; with the map down to half a screen the argument
for giving it all to the map is stronger than it was, so this stays as it is
until someone says otherwise.

The sheet is a fixed half with no grab handle and no slide-in. A handle that
does not drag is a lie about what the sheet does, and the map resizes in the
same moment the sheet would slide, so animating one without choreographing the
other would look worse than neither.

Not yet run on a device.
